é pochibile

+++ SRC/built-ins/utils2.c
@@ -6,7 +6,7 @@
 /*   By: jsebasti <jsebasti@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/05 00:09:50 by jsebasti          #+#    #+#             */
-/*   Updated: 2023/06/18 13:36:35 by jsebasti         ###   ########.fr       */
+/*   Updated: 2023/06/27 12:17:22 by jsebasti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,10 +41,10 @@ void	new_err(char *s1, char *s2, char *s3)
 	ft_putstr_fd(s3, STDERR);
 }
 
-long long	ft_atoll(char *str)
+long int	ft_atoll(char *str)
 {
-	long long	neg;
-	long long	sum;
+	long int	neg;
+	long int	sum;
 
 	neg = 1;
 	sum = 0;
